+++ src/components/setting/sync/DropboxAuth.ts
@@ -1,4 +1,4 @@
-import { Dropbox as _Dropbox, files } from 'dropbox';
+import { Dropbox as _Dropbox } from 'dropbox';
 import { authorize } from 'react-native-app-auth';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore: dropbox didnt have fileBlob in their sdk anywhere but UPGRADING.md
@@ -9,9 +9,11 @@ import { getArrayBufferForBlob } from 'react-native-blob-jsi-helper';
 import { DROPBOX_KEY, DROPBOX_SECRET } from '@env';
 import { logger } from '@utils/Logger';
 import GenericSyncButton, { GenericProps } from './GenericSyncButton';
-
-const DEFAULT_FILE_NAME = 'nox.noxBackup';
-const DEFAULT_FILE_PATH = `/${DEFAULT_FILE_NAME}`;
+import {
+  checkAuthentication,
+  noxBackup,
+  noxRestore,
+} from '@utils/sync/Dropbox';
 
 const config = {
   clientId: DROPBOX_KEY,
@@ -42,7 +44,7 @@ let dbx = new _Dropbox({
  * @param {function} errorHandling
  */
 export const getAuth = async (
-  callback = () => checkAuthentication().then(console.log),
+  callback = () => checkAuthentication(dbx).then(console.log),
   errorHandling = logger.error
 ) => {
   const authState = await authorize(config);
@@ -54,97 +56,6 @@ export const getAuth = async (
     callback();
   } else {
     errorHandling('no response url returned. auth aborted by user.');
-  }
-};
-
-/**
- * lists the noxplayer setting in dropbox.
- * returns either null if nothing is found, or the path_display of it
- * that can be used to retrieve content.
- * @param {string} query
- * @returns {string}
- */
-
-const find = async (query = DEFAULT_FILE_NAME) => {
-  const data = await dbx.filesSearchV2({
-    query,
-    options: {
-      order_by: { '.tag': 'last_modified_time' },
-    },
-  });
-  try {
-    const fileMetadata = data.result.matches[0]
-      .metadata as files.MetadataV2Metadata;
-    return fileMetadata.metadata.path_display;
-  } catch (e) {
-    logger.warn(`no ${query} found.`);
-    return null;
-  }
-};
-
-/**
- * upload the noxplayer setting file to dropbox, with the mode
- * overwrite. As a sync function there is no need to keep multiple
- * versions.
- * @param {Object} content
- * @param {string} fpath
- * @returns
- */
-const upload = async (content: Uint8Array, fpath = DEFAULT_FILE_PATH) => {
-  return await dbx.filesUpload({
-    path: fpath,
-    mode: { '.tag': 'overwrite' },
-    contents: content,
-  });
-};
-// upload({'new': 'content'}).then(console.log)
-
-/**
- * download the noxplayer setting from dropbox.
- * returns the parsed JSON object or null if not found.
- * @param {string} fpath
- * @returns playerSetting object, or null
- */
-const download = async (fpath = DEFAULT_FILE_PATH) => {
-  if (fpath === null) {
-    return null;
-  }
-  const downloadedFile = await dbx.filesDownload({ path: fpath });
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore: dropbox didnt have fileBlob in their sdk anywhere but UPGRADING.md
-  const blob = getArrayBufferForBlob(downloadedFile.result.fileBlob);
-  return new Uint8Array(blob);
-};
-
-/**
- * wraps up find noxplayer setting and download in one function;
- * returns the JSON object of settting or null if not found.
- * @returns playerSetting object, or null
- */
-export const noxRestore = async () => {
-  const noxFile = await find();
-  if (!noxFile) {
-    throw new Error('noxfile is not found on dropbox.');
-  }
-  return await download(noxFile);
-};
-
-/**
- * wraps up upload noxplayer setting. returns the response
- * if successful.
- * @param {Object} content
- * @returns
- */
-export const noxBackup = async (content: Uint8Array) => {
-  return await upload(content);
-};
-
-const checkAuthentication = async () => {
-  try {
-    await dbx.usersGetCurrentAccount();
-    return true;
-  } catch (e) {
-    return false;
   }
 };
 
@@ -163,7 +74,7 @@ export const loginDropbox = async (
   errorCallback = logger.error
 ) => {
   try {
-    if (!(await checkAuthentication())) {
+    if (!(await checkAuthentication(dbx))) {
       logger.debug('dropbox token expired, need to log in');
       await getAuth(callback, errorCallback);
     } else {
@@ -179,8 +90,8 @@ export const loginDropbox = async (
 const DropboxSyncButton = ({ restoreFromUint8Array }: GenericProps) =>
   GenericSyncButton({
     restoreFromUint8Array,
-    noxBackup,
-    noxRestore,
+    noxBackup: v => noxBackup(dbx, v),
+    noxRestore: () => noxRestore(dbx, async v => getArrayBufferForBlob(v)),
     login: loginDropbox,
   });
